toolkits/websearch/pyproject.toml
[project]
name = "websearch"
version = "0.1.0"
description = "Web search toolkit"
requires-python = ">=3.10"
dependencies = [
    "tavily-python>=0.5.0",
    "unifai-sdk>=0.2.1",
]
